Rename mapGenerator.py to old_mapGenerator.py
